Moved instruction decoder into its own source file
