+++ pioemu/emulation.py
@@ -13,35 +13,10 @@
 # limitations under the License.
 from dataclasses import replace
 from enum import Enum, unique
-from functools import partial, wraps
+from .instruction_decoder import InstructionDecoder
 from .state import State
-from .conditions import (
-    x_register_equals_zero,
-    x_register_not_equal_to_y_register,
-    x_register_not_equal_to_zero,
-    y_register_equals_zero,
-    y_register_not_equal_to_zero,
-)
 from .instructions import (
-    jmp,
-    jmp_when_x_is_non_zero_and_post_decrement,
-    jmp_when_y_is_non_zero_and_post_decrement,
-    mov_into_isr,
-    mov_into_osr,
-    mov_into_pins,
-    mov_into_x,
-    mov_into_y,
-    out_null,
-    out_pindirs,
-    out_pins,
-    out_x,
-    out_y,
     pull_blocking,
-    pull_nonblocking,
-    set_pins,
-    set_pindirs,
-    set_x,
-    set_y,
     wait_for_gpio_low,
     wait_for_gpio_high,
 )
@@ -70,9 +45,9 @@ def emulate(
     stop_condition = stop_condition or (lambda state: False)
 
     if shift_osr_right:
-        instruction_lookup = map_opcodes_to_callables(shift_right)
+        instruction_decoder = InstructionDecoder(shift_right)
     else:
-        instruction_lookup = map_opcodes_to_callables(shift_left)
+        instruction_decoder = InstructionDecoder(shift_left)
 
     current_state = initial_state
 
@@ -89,7 +64,7 @@ def emulate(
         else:
             data_field = opcode & 0x1F
 
-        instruction = instruction_lookup.get(opcode & 0xE0E0, None)
+        instruction = instruction_decoder.decode(opcode)
 
         if instruction is None:
             return
@@ -128,17 +103,6 @@ def _apply_side_set_to_pin_values(state, pin_base, pin_count, pin_values):
     return replace(state, pin_values=new_pin_values)
 
 
-def _normalize_bit_count(function):
-    @wraps(function)
-    def wrapper(bit_count, state):
-        if bit_count == 0:
-            return function(32, state)
-        else:
-            return function(bit_count, state)
-
-    return wrapper
-
-
 def _read_source(source, state):
     # TODO: Consider using the new match statement when widely available
     if source == MoveSource.PINS:
@@ -164,32 +128,3 @@ def is_instruction_stalled(previous_state, current_state, instruction):
         return current_state.program_counter == previous_state.program_counter
     else:
         return False
-
-
-def map_opcodes_to_callables(shifter_for_osr):
-    return {
-        0x0000: partial(jmp, lambda state: True),
-        0x0020: partial(jmp, x_register_equals_zero),
-        0x0040: jmp_when_x_is_non_zero_and_post_decrement,
-        0x0060: partial(jmp, y_register_equals_zero),
-        0x0080: jmp_when_y_is_non_zero_and_post_decrement,
-        0x00A0: partial(jmp, x_register_not_equal_to_y_register),
-        0x2020: wait_for_gpio_low,
-        0x2080: wait_for_gpio_high,
-        0x6000: _normalize_bit_count(partial(out_pins, shifter_for_osr)),
-        0x6020: _normalize_bit_count(partial(out_x, shifter_for_osr)),
-        0x6040: _normalize_bit_count(partial(out_y, shifter_for_osr)),
-        0x6060: _normalize_bit_count(partial(out_null, shifter_for_osr)),
-        0x6080: _normalize_bit_count(partial(out_pindirs, shifter_for_osr)),
-        0x8080: pull_nonblocking,
-        0x80A0: pull_blocking,
-        0xA000: mov_into_pins,
-        0xA020: mov_into_x,
-        0xA040: mov_into_y,
-        0xA0C0: mov_into_isr,
-        0xA0E0: mov_into_osr,
-        0xE080: set_pindirs,
-        0xE000: set_pins,
-        0xE020: set_x,
-        0xE040: set_y,
-    }
